% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinRollup.R
\name{protein_rollup}
\alias{protein_rollup}
\title{Executes the protein rollup}
\usage{
protein_rollup(protein_ids, pep_mat, rollup_func = rrollup,
  protein_col_name = "Protein", get_debug_info = FALSE,
  one_hit_wonders = FALSE, min_presence = 0, min_overlap = 3,
  debug_protein = NA)
}
\arguments{
\item{protein_ids}{Vector containing protein IDs corresponding to each peptide}

\item{pep_mat}{Matrix containing peptide expression values}

\item{rollup_func}{What rollup function to apply}

\item{protein_col_name}{Name of protein column in input}

\item{get_debug_info}{Print debug information on the go}

\item{one_hit_wonders}{Whether proteins based in a single peptide should be retained}

\item{min_presence}{Value between 0 and 1. What fraction of values must be present in the most present peptide for the protein to be valid}

\item{min_overlap}{The number of non-missing values required from each peptide}

\item{debug_protein}{Send in a protein ID to print more extensive debug information}
}
\value{
Produced protein matrix
}
\description{
Executes the protein rollup
}
